requirements.txt
flask
flask-cors
faiss-cpu
openai
sentence-transformers
